Splitting to the ILD specific utils
